v0.1

+++ fast_train_hat.py
@@ -193,7 +193,7 @@ def trades_loss(logits,
     criterion_kl = nn.KLDivLoss(size_average=False)
     loss_natural = F.cross_entropy(logits, targets)
     loss_robust = (1.0 / logits.shape[0]) * criterion_kl(F.log_softmax(logits_adv, dim=1), F.softmax(logits, dim=1))
-    loss = loss_natural + float(2) * loss_robust
+    loss = loss_natural + float(5) * loss_robust
     return loss
 
 def main_worker(rank, ngpus_per_node=ngpus_per_node):
@@ -261,6 +261,7 @@ def main_worker(rank, ngpus_per_node=ngpus_per_node):
         rprint('\nEpoch: %d' % (epoch+1), rank)
         train(net, std, trainloader, optimizer, lr_scheduler, scaler, attack)
         test(net, testloader, attack, rank)
+
 
 
 def run():
